Delete redundant tests
